Heavy refactoring
Refactored the `folio` method and simplified it. Added an example and
added it to documentation. Also began a method to add ads based on
previous block element methods. Added a temporary header to test the
folio.

Lightly refactored the `simpleFooter` method to remove clutter. Also
added example and documentation.

Shortened the amount of grass in the data.html file.