Modified read me, added a query plugin version
I modified readme.md to make mention of Jay’s fine work. Added a
Sinclair in the wild section with links to projects making use of it. I
copied over a version of the repo prior to the merge which leaves the
simpler version in case we want to use it as a jQuery plugin again.